+++ src/sh_pwd.c
@@ -6,7 +6,7 @@
 /*   By: syamashi <syamashi@student.42.tokyo>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/15 15:22:52 by syamashi          #+#    #+#             */
-/*   Updated: 2021/02/15 15:24:02 by syamashi         ###   ########.fr       */
+/*   Updated: 2021/02/15 20:41:22 by syamashi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
